+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Changes on `main` will be listed here.
 
 ### Features
 
--   Added `confirmingDelete` flag to `Row`, which changes the delete icon to be more pronouced.
+-   [[#12](https://github.com/diatche/react-native-form-model/pull/12)] Added `confirmingDelete` flag to `Row`, which changes the delete icon to be more pronouced.
 
 ### Breaking Changes